+++ android/shared/src/commonMain/kotlin/com/yeobaek/data/model/CommentModel.kt
@@ -1,7 +1,7 @@
 package com.yeobaek.data.model
 
 data class CommentModel(
-    val commentId: Int,
+    val commentId: Long,
     val memberId: Int,
     val nickname: String,
     val content: String,
